Split platform tests and add new tag `test_functions_kube`
