+++ pkg/platform/kube/test/function/function_test.go
@@ -1,4 +1,4 @@
-//go:build test_integration && test_kube
+//go:build test_integration && test_kube && test_functions_kube
 
 /*
 Copyright 2023 The Nuclio Authors.
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package test
+package function
 
 import (
 	"bytes"
@@ -25,7 +25,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"net/http"
 	"os"
 	"path"
 	"sort"
@@ -38,7 +37,7 @@ import (
 	"github.com/nuclio/nuclio/pkg/platform"
 	"github.com/nuclio/nuclio/pkg/platform/kube"
 	nuclioio "github.com/nuclio/nuclio/pkg/platform/kube/apis/nuclio.io/v1beta1"
-	"github.com/nuclio/nuclio/pkg/platform/kube/ingress"
+	suite2 "github.com/nuclio/nuclio/pkg/platform/kube/test/suite"
 	"github.com/nuclio/nuclio/pkg/platformconfig"
 	"github.com/nuclio/nuclio/pkg/processor/build"
 	"github.com/nuclio/nuclio/pkg/processor/build/runtimeconfig"
@@ -60,7 +59,7 @@ import (
 )
 
 type DeployFunctionTestSuite struct {
-	KubeTestSuite
+	suite2.KubeTestSuite
 }
 
 func (suite *DeployFunctionTestSuite) TestDeployCronTriggerK8sWithJSONEventBody() {
@@ -281,7 +280,7 @@ func (suite *DeployFunctionTestSuite) TestVolumeOnceMountTwice() {
 	suite.DeployFunction(createFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
 		podName := fmt.Sprintf("deployment/%s", kube.DeploymentNameFromFunctionName(functionName))
 		for _, mountPath := range mountPaths {
-			results, err := suite.executeKubectl([]string{
+			results, err := suite.ExecuteKubectl([]string{
 				"exec",
 				podName,
 				"--",
@@ -460,7 +459,7 @@ func (suite *DeployFunctionTestSuite) TestSecurityContext() {
 
 		// verify running function indeed using the right uid / gid / groups
 		podName := fmt.Sprintf("deployment/%s", kube.DeploymentNameFromFunctionName(functionName))
-		results, err := suite.executeKubectl([]string{"exec", podName, "--", "id"}, nil)
+		results, err := suite.ExecuteKubectl([]string{"exec", podName, "--", "id"}, nil)
 		suite.Require().NoError(err, "Failed to execute `id` command on function pod")
 		suite.Require().Equal(fmt.Sprintf(`uid=%d gid=%d groups=%d,%d`,
 			runAsUserID,
@@ -642,9 +641,9 @@ func (suite *DeployFunctionTestSuite) TestDefaultHTTPTrigger() {
 	suite.DeployFunction(createDefaultTriggerFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
 
 		// ensure only 1 http trigger exists, always.
-		suite.ensureTriggerAmount(defaultTriggerFunctionName, "http", 1)
+		suite.EnsureTriggerAmount(defaultTriggerFunctionName, "http", 1)
 		defaultHTTPTrigger := functionconfig.GetDefaultHTTPTrigger()
-		return suite.verifyCreatedTrigger(defaultTriggerFunctionName, defaultHTTPTrigger)
+		return suite.VerifyCreatedTrigger(defaultTriggerFunctionName, defaultHTTPTrigger)
 	})
 
 	customTriggerFunctionName := "custom-http-trigger"
@@ -660,8 +659,8 @@ func (suite *DeployFunctionTestSuite) TestDefaultHTTPTrigger() {
 	suite.DeployFunction(createCustomTriggerFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
 
 		// ensure only 1 http trigger exists, always.
-		suite.ensureTriggerAmount(customTriggerFunctionName, "http", 1)
-		return suite.verifyCreatedTrigger(customTriggerFunctionName, customTrigger)
+		suite.EnsureTriggerAmount(customTriggerFunctionName, "http", 1)
+		return suite.VerifyCreatedTrigger(customTriggerFunctionName, customTrigger)
 	})
 }
 
@@ -1181,6 +1180,7 @@ func (suite *DeployFunctionTestSuite) TestCleanFlexVolumeSubPath() {
 			},
 		},
 	}
+	createFunctionOptions.FunctionConfig.Spec.ReadinessTimeoutSeconds = 10
 
 	// deploy function, we expect a failure due to the v3io access key being a dummy value
 	_, err := suite.DeployFunctionExpectError(createFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
@@ -1742,7 +1742,7 @@ func (suite *DeployFunctionTestSuite) validatePodLogsContainData(podName string,
 }
 
 type DeleteFunctionTestSuite struct {
-	KubeTestSuite
+	suite2.KubeTestSuite
 }
 
 func (suite *DeleteFunctionTestSuite) TestDeleteFunctionWhichHasApiGateway() {
@@ -1900,7 +1900,7 @@ def init_context(context):
 }
 
 type UpdateFunctionTestSuite struct {
-	KubeTestSuite
+	suite2.KubeTestSuite
 }
 
 func (suite *UpdateFunctionTestSuite) TestSanity() {
@@ -2019,528 +2019,11 @@ func (suite *UpdateFunctionTestSuite) TestUpdateFunctionWithSecret() {
 	suite.Require().Equal(password, secretData[secretPasswordKey])
 }
 
-type DeployAPIGatewayTestSuite struct {
-	KubeTestSuite
-}
-
-func (suite *DeployAPIGatewayTestSuite) TestDexAuthMode() {
-	functionName := "some-function-name"
-	apiGatewayName := "some-api-gateway-name"
-	createFunctionOptions := suite.CompileCreateFunctionOptions(functionName)
-	configOauth2ProxyURL := "config-oauth2-url"
-	suite.PlatformConfiguration.IngressConfig = platformconfig.IngressConfig{
-		Oauth2ProxyURL: configOauth2ProxyURL,
-	}
-	suite.DeployFunction(createFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
-		createAPIGatewayOptions := suite.CompileCreateAPIGatewayOptions(apiGatewayName, functionName)
-		createAPIGatewayOptions.APIGatewayConfig.Spec.AuthenticationMode = ingress.AuthenticationModeOauth2
-		err := suite.DeployAPIGateway(createAPIGatewayOptions, func(ingress *networkingv1.Ingress) {
-			suite.Require().NotContains(ingress.Annotations, "nginx.ingress.kubernetes.io/auth-signin")
-			suite.Require().Contains(ingress.Annotations["nginx.ingress.kubernetes.io/auth-url"], configOauth2ProxyURL)
-			suite.Require().Contains(ingress.Annotations["nginx.ingress.kubernetes.io/configuration-snippet"],
-				fmt.Sprintf(`proxy_set_header X-Nuclio-Target "%s";`, functionName))
-		})
-		suite.Require().NoError(err)
-
-		overrideOauth2ProxyURL := "override-oauth2-url"
-		createAPIGatewayOptions = suite.CompileCreateAPIGatewayOptions(apiGatewayName, functionName)
-		createAPIGatewayOptions.APIGatewayConfig.Spec.AuthenticationMode = ingress.AuthenticationModeOauth2
-		createAPIGatewayOptions.APIGatewayConfig.Spec.Authentication = &platform.APIGatewayAuthenticationSpec{
-			DexAuth: &ingress.DexAuth{
-				Oauth2ProxyURL:               overrideOauth2ProxyURL,
-				RedirectUnauthorizedToSignIn: true,
-			},
-		}
-		err = suite.DeployAPIGateway(createAPIGatewayOptions, func(ingress *networkingv1.Ingress) {
-			suite.Assert().Contains(ingress.Annotations, "nginx.ingress.kubernetes.io/auth-signin")
-			suite.Assert().Contains(ingress.Annotations["nginx.ingress.kubernetes.io/auth-signin"], overrideOauth2ProxyURL)
-			suite.Assert().Contains(ingress.Annotations["nginx.ingress.kubernetes.io/auth-url"], overrideOauth2ProxyURL)
-		})
-		suite.Require().NoError(err)
-
-		return true
-	})
-}
-
-func (suite *DeployAPIGatewayTestSuite) TestFunctionWithTwoGateways() {
-	functionName := "some-function-name"
-	apiGatewayName1 := "api-gateway-1"
-	apiGatewayName2 := "api-gateway-2"
-	createFunctionOptions := suite.CompileCreateFunctionOptions(functionName)
-
-	suite.DeployFunction(createFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
-		// create first api gateway on top of given function
-		createAPIGatewayOptions1 := suite.CompileCreateAPIGatewayOptions(apiGatewayName1, functionName)
-		createAPIGatewayOptions1.APIGatewayConfig.Spec.AuthenticationMode = ingress.AuthenticationModeNone
-		createAPIGatewayOptions1.APIGatewayConfig.Spec.Host = "host1.com"
-
-		err := suite.DeployAPIGateway(createAPIGatewayOptions1, func(ingressObj *networkingv1.Ingress) {
-			// create second api gateway on top of the same function
-			createAPIGatewayOptions2 := suite.CompileCreateAPIGatewayOptions(apiGatewayName2, functionName)
-			createAPIGatewayOptions2.APIGatewayConfig.Spec.AuthenticationMode = ingress.AuthenticationModeNone
-			createAPIGatewayOptions2.APIGatewayConfig.Spec.Host = "host2.com"
-
-			err := suite.DeployAPIGateway(createAPIGatewayOptions2, func(ingress *networkingv1.Ingress) {
-				// check that both gateways are invokable
-				_, err := http.Get(fmt.Sprintf("http://%s", createAPIGatewayOptions1.APIGatewayConfig.Spec.Host))
-				suite.Require().NoError(err)
-
-				_, err = http.Get(fmt.Sprintf("http://%s", createAPIGatewayOptions2.APIGatewayConfig.Spec.Host))
-				suite.Require().NoError(err)
-			})
-			suite.Require().NoError(err)
-		})
-		suite.Require().NoError(err)
-
-		return true
-	})
-}
-
-func (suite *DeployAPIGatewayTestSuite) TestUpdate() {
-	projectName := "some-project-" + xid.New().String()
-	functionName := "function-name-" + xid.New().String()
-	apiGatewayName := "apigw-name-" + xid.New().String()
-	createFunctionOptions := suite.CompileCreateFunctionOptions(functionName)
-
-	// create project
-	err := suite.Platform.CreateProject(suite.Ctx, &platform.CreateProjectOptions{
-		ProjectConfig: &platform.ProjectConfig{
-			Meta: platform.ProjectMeta{
-				Name:      projectName,
-				Namespace: suite.Namespace,
-			},
-		},
-	})
-	suite.Require().NoError(err, "Failed to create project")
-	createFunctionOptions.FunctionConfig.Meta.Labels[common.NuclioResourceLabelKeyProjectName] = projectName
-
-	suite.DeployFunction(createFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
-		createAPIGatewayOptions := suite.CompileCreateAPIGatewayOptions(apiGatewayName, functionName)
-		beforeUpdateHostValue := "before-update-host.com"
-		createAPIGatewayOptions.APIGatewayConfig.Spec.Host = beforeUpdateHostValue
-		createAPIGatewayOptions.APIGatewayConfig.Meta.Labels[common.NuclioResourceLabelKeyProjectName] = projectName
-		createAPIGatewayOptions.APIGatewayConfig.Meta.Annotations = map[string]string{
-			"some/annotation": "some-value",
-		}
-
-		// create
-		err := suite.Platform.CreateAPIGateway(suite.Ctx, createAPIGatewayOptions)
-		suite.Require().NoError(err)
-
-		// delete leftovers
-		defer suite.Platform.DeleteAPIGateway(suite.Ctx, &platform.DeleteAPIGatewayOptions{ // nolint: errcheck
-			Meta: createAPIGatewayOptions.APIGatewayConfig.Meta,
-		})
-
-		suite.WaitForAPIGatewayState(&platform.GetAPIGatewaysOptions{
-			Name:      createAPIGatewayOptions.APIGatewayConfig.Meta.Name,
-			Namespace: createAPIGatewayOptions.APIGatewayConfig.Meta.Namespace,
-		}, platform.APIGatewayStateReady, 10*time.Second)
-
-		ingressInstance := suite.GetAPIGatewayIngress(createAPIGatewayOptions.APIGatewayConfig.Meta.Name, false)
-		suite.Require().Equal(beforeUpdateHostValue, ingressInstance.Spec.Rules[0].Host)
-
-		// ensure ingress labels were created correctly
-		suite.Require().Equal("apigateway", ingressInstance.Labels[common.NuclioLabelKeyClass])
-		suite.Require().Equal("ingress-manager", ingressInstance.Labels[common.NuclioLabelKeyApp])
-		suite.Require().Equal(apiGatewayName, ingressInstance.Labels[common.NuclioResourceLabelKeyApiGatewayName])
-		suite.Require().Equal(projectName, ingressInstance.Labels[common.NuclioResourceLabelKeyProjectName])
-		suite.Require().Equal("some-value", ingressInstance.Annotations["some/annotation"])
-
-		// change host, update
-		afterUpdateHostValue := "after-update-host.com"
-		annotations := map[string]string{
-			"annotation-key": "annotation-value",
-		}
-		createAPIGatewayOptions.APIGatewayConfig.Spec.Host = afterUpdateHostValue
-		createAPIGatewayOptions.APIGatewayConfig.Meta.Annotations = annotations
-		err = suite.Platform.UpdateAPIGateway(suite.Ctx, &platform.UpdateAPIGatewayOptions{
-			APIGatewayConfig: createAPIGatewayOptions.APIGatewayConfig,
-		})
-		suite.Require().NoError(err)
-
-		getAPIGatewayOptions := &platform.GetAPIGatewaysOptions{
-			Name:      createAPIGatewayOptions.APIGatewayConfig.Meta.Name,
-			Namespace: createAPIGatewayOptions.APIGatewayConfig.Meta.Namespace,
-		}
-		suite.WaitForAPIGatewayState(getAPIGatewayOptions, platform.APIGatewayStateReady, 10*time.Second)
-
-		ingressInstance = suite.GetAPIGatewayIngress(createAPIGatewayOptions.APIGatewayConfig.Meta.Name, false)
-		suite.Require().Equal(afterUpdateHostValue, ingressInstance.Spec.Rules[0].Host)
-
-		apiGateway := suite.GetAPIGateway(getAPIGatewayOptions)
-		suite.Require().Equal(annotations, apiGateway.GetConfig().Meta.Annotations)
-		return true
-	})
-}
-
-func (suite *DeployAPIGatewayTestSuite) TestSetSpecificPort() {
-	functionName := "some-function-name"
-	apiGatewayName := "api-gateway-1"
-	sidecarPort := 8050
-	createFunctionOptions := suite.CompileCreateFunctionOptions(functionName)
-	createFunctionOptions.FunctionConfig.Spec.Sidecars = []*v1.Container{
-		{
-			Name:    "sidecar",
-			Image:   "busybox",
-			Command: []string{"sh", "-c", "while true; do echo 'sidecar'; sleep 1; done"},
-			Ports: []v1.ContainerPort{
-				{
-					Name:          "sidecar-port",
-					ContainerPort: int32(sidecarPort),
-				},
-			},
-		},
-	}
-
-	suite.DeployFunction(createFunctionOptions, func(deployResult *platform.CreateFunctionResult) bool {
-		// create an api gateway on top of the function with a specific port to the sidecar
-		createAPIGatewayOptions1 := suite.CompileCreateAPIGatewayOptions(apiGatewayName, functionName)
-		createAPIGatewayOptions1.APIGatewayConfig.Spec.AuthenticationMode = ingress.AuthenticationModeNone
-		createAPIGatewayOptions1.APIGatewayConfig.Spec.Host = "host1.com"
-		createAPIGatewayOptions1.APIGatewayConfig.Spec.Upstreams[0].Port = sidecarPort
-
-		err := suite.DeployAPIGateway(createAPIGatewayOptions1, func(ingressObj *networkingv1.Ingress) {
-			// check that the ingress has the correct port
-			suite.Require().Equal(int32(sidecarPort), ingressObj.Spec.Rules[0].HTTP.Paths[0].Backend.Service.Port.Number)
-		})
-		suite.Require().NoError(err)
-
-		return true
-	})
-}
-
-type ProjectTestSuite struct {
-	KubeTestSuite
-}
-
-func (suite *ProjectTestSuite) TestCreate() {
-	projectConfig := platform.ProjectConfig{
-		Meta: platform.ProjectMeta{
-			Name:      "test-project",
-			Namespace: suite.Namespace,
-			Labels: map[string]string{
-				"label-key": "label-value",
-			},
-			Annotations: map[string]string{
-				"annotation-key": "annotation-value",
-			},
-		},
-		Spec: platform.ProjectSpec{
-			Description: "some description",
-		},
-	}
-
-	// create project
-	err := suite.Platform.CreateProject(suite.Ctx, &platform.CreateProjectOptions{
-		ProjectConfig: &projectConfig,
-	})
-	suite.Require().NoError(err, "Failed to create project")
-	defer func() {
-		err = suite.Platform.DeleteProject(suite.Ctx, &platform.DeleteProjectOptions{
-			Meta:     projectConfig.Meta,
-			Strategy: platform.DeleteProjectStrategyRestricted,
-		})
-		suite.Require().NoError(err, "Failed to delete project")
-	}()
-
-	// get created project
-	projects, err := suite.Platform.GetProjects(suite.Ctx, &platform.GetProjectsOptions{
-		Meta: projectConfig.Meta,
-	})
-	suite.Require().NoError(err, "Failed to get projects")
-	suite.Require().Equal(len(projects), 1)
-
-	// requested and created project are equal
-	createdProject := projects[0]
-	suite.Require().Equal(projectConfig, *createdProject.GetConfig())
-}
-
-func (suite *ProjectTestSuite) TestCreateFromLeaderIgnoreInvalidLabels() {
-	invalidLabelKey1 := "invalid.label@-key"
-	invalidLabelKey2 := "the-key-is-invalid"
-	projectConfig := platform.ProjectConfig{
-		Meta: platform.ProjectMeta{
-			Name:      "test-project",
-			Namespace: suite.Namespace,
-			Labels: map[string]string{
-				invalidLabelKey1: "label-value",
-				invalidLabelKey2: "inva!id_v8$alue",
-			},
-		},
-		Spec: platform.ProjectSpec{
-			Description: "some description",
-		},
-	}
-
-	// set the platform's leader kind
-	suite.Platform.GetConfig().ProjectsLeader = &platformconfig.ProjectsLeader{
-		Kind: platformconfig.ProjectsLeaderKindMock,
-	}
-
-	// create project
-	err := suite.Platform.CreateProject(suite.Ctx, &platform.CreateProjectOptions{
-		ProjectConfig: &projectConfig,
-		RequestOrigin: platformconfig.ProjectsLeaderKindMock,
-	})
-	suite.Require().NoError(err, "Failed to create project")
-	defer func() {
-		err = suite.Platform.DeleteProject(suite.Ctx, &platform.DeleteProjectOptions{
-			Meta:     projectConfig.Meta,
-			Strategy: platform.DeleteProjectStrategyRestricted,
-		})
-		suite.Require().NoError(err, "Failed to delete project")
-	}()
-
-	// get created project
-	projects, err := suite.Platform.GetProjects(suite.Ctx, &platform.GetProjectsOptions{
-		Meta: projectConfig.Meta,
-	})
-	suite.Require().NoError(err, "Failed to get projects")
-	suite.Require().Equal(len(projects), 1)
-
-	// verify created project does not contain the invalid label
-	createdProject := projects[0]
-	suite.Require().NotContains(createdProject.GetConfig().Meta.Labels, invalidLabelKey1)
-	suite.Require().NotContains(createdProject.GetConfig().Meta.Labels, invalidLabelKey2)
-}
-
-func (suite *ProjectTestSuite) TestUpdate() {
-	projectConfig := platform.ProjectConfig{
-		Meta: platform.ProjectMeta{
-			Name:      "test-project",
-			Namespace: suite.Namespace,
-			Labels: map[string]string{
-				"something": "here",
-			},
-			Annotations: map[string]string{
-				"annotation-key": "annotation-value",
-			},
-		},
-		Spec: platform.ProjectSpec{
-			Description: "Simple description",
-		},
-	}
-
-	// create project
-	err := suite.Platform.CreateProject(suite.Ctx, &platform.CreateProjectOptions{
-		ProjectConfig: &projectConfig,
-	})
-	suite.Require().NoError(err, "Failed to create project")
-
-	// delete leftover
-	defer func() {
-		err := suite.Platform.DeleteProject(suite.Ctx, &platform.DeleteProjectOptions{
-			Meta:     projectConfig.Meta,
-			Strategy: platform.DeleteProjectStrategyRestricted,
-		})
-		suite.Require().NoError(err, "Failed to delete project")
-	}()
-
-	// change project annotations
-	projectConfig.Meta.Annotations["annotation-key"] = "annotation-value-changed"
-	projectConfig.Meta.Annotations["added-annotation"] = "added-annotation-value"
-
-	// change project labels
-	projectConfig.Meta.Labels["label-key"] = "label-value-changed"
-	projectConfig.Meta.Labels["added-label"] = "added-label-value"
-
-	// update project
-	err = suite.Platform.UpdateProject(suite.Ctx, &platform.UpdateProjectOptions{
-		ProjectConfig: projectConfig,
-	})
-	suite.Require().NoError(err, "Failed to update project")
-
-	// get updated project
-	updatedProject := suite.GetProject(&platform.GetProjectsOptions{
-		Meta: projectConfig.Meta,
-	})
-	suite.Require().Empty(cmp.Diff(projectConfig, *updatedProject.GetConfig(),
-		cmp.Options{
-			cmpopts.IgnoreFields(projectConfig.Status, "UpdatedAt"), // automatically populated
-		}))
-}
-
-func (suite *ProjectTestSuite) TestDelete() {
-	projectConfig := platform.ProjectConfig{
-		Meta: platform.ProjectMeta{
-			Name:      "test-project",
-			Namespace: suite.Namespace,
-			Labels: map[string]string{
-				"something": "here",
-			},
-			Annotations: map[string]string{
-				"annotation-key": "annotation-value",
-			},
-		},
-		Spec: platform.ProjectSpec{
-			Description: "Simple description",
-		},
-	}
-
-	// create project
-	err := suite.Platform.CreateProject(suite.Ctx, &platform.CreateProjectOptions{
-		ProjectConfig: &projectConfig,
-	})
-	suite.Require().NoError(err, "Failed to create project")
-
-	// delete project
-	err = suite.Platform.DeleteProject(suite.Ctx, &platform.DeleteProjectOptions{
-		Meta:     projectConfig.Meta,
-		Strategy: platform.DeleteProjectStrategyRestricted,
-	})
-	suite.Require().NoError(err, "Failed to delete project")
-
-	// ensure project does not exist
-	projects, err := suite.Platform.GetProjects(suite.Ctx, &platform.GetProjectsOptions{
-		Meta: projectConfig.Meta,
-	})
-	suite.Require().NoError(err, "Failed to get projects")
-	suite.Require().Equal(0, len(projects))
-}
-
-func (suite *ProjectTestSuite) TestDeleteCascading() {
-
-	// create project
-	projectToDeleteConfig := platform.ProjectConfig{
-		Meta: platform.ProjectMeta{
-			Name:      "project-to-delete",
-			Namespace: suite.Namespace,
-		},
-	}
-	err := suite.Platform.CreateProject(suite.Ctx, &platform.CreateProjectOptions{
-		ProjectConfig: &projectToDeleteConfig,
-	})
-	suite.Require().NoError(err, "Failed to create project")
-
-	// create 2 functions (deleted along with `projectToDeleteConfig`)
-
-	// create function A
-	functionToDeleteA := suite.CreateImportedFunction("func-to-delete-a", projectToDeleteConfig.Meta.Name)
-	functionToDeleteB := suite.CreateImportedFunction("func-to-delete-b", projectToDeleteConfig.Meta.Name)
-
-	// delete leftovers
-	defer suite.Platform.DeleteFunction(suite.Ctx, &platform.DeleteFunctionOptions{ // nolint: errcheck
-		FunctionConfig: *functionToDeleteA,
-	})
-	defer suite.Platform.DeleteFunction(suite.Ctx, &platform.DeleteFunctionOptions{ // nolint: errcheck
-		FunctionConfig: *functionToDeleteB,
-	})
-
-	// create api gateway for function A (deleted along with `projectToDeleteConfig`)
-	createAPIGatewayOptions := suite.CompileCreateAPIGatewayOptions("apigw-to-delete",
-		functionToDeleteA.Meta.Name)
-	createAPIGatewayOptions.APIGatewayConfig.Meta.Labels[common.NuclioResourceLabelKeyProjectName] = projectToDeleteConfig.Meta.Name
-	err = suite.Platform.CreateAPIGateway(suite.Ctx, createAPIGatewayOptions)
-	suite.Require().NoError(err)
-	defer suite.Platform.DeleteAPIGateway(suite.Ctx, &platform.DeleteAPIGatewayOptions{ // nolint: errcheck
-		Meta: createAPIGatewayOptions.APIGatewayConfig.Meta,
-	})
-
-	suite.WaitForAPIGatewayState(&platform.GetAPIGatewaysOptions{
-		Name:      createAPIGatewayOptions.APIGatewayConfig.Meta.Name,
-		Namespace: createAPIGatewayOptions.APIGatewayConfig.Meta.Namespace,
-	}, platform.APIGatewayStateReady, 10*time.Second)
-
-	// create 2 function events for function B (deleted along with `projectToDeleteConfig`)
-	functionEventA := suite.CompileCreateFunctionEventOptions("function-event-a", functionToDeleteB.Meta.Name)
-	err = suite.Platform.CreateFunctionEvent(suite.Ctx, functionEventA)
-	suite.Require().NoError(err)
-	defer suite.Platform.DeleteFunctionEvent(suite.Ctx, &platform.DeleteFunctionEventOptions{ // nolint: errcheck
-		Meta: platform.FunctionEventMeta{
-			Name:      functionEventA.FunctionEventConfig.Meta.Name,
-			Namespace: suite.Namespace,
-		},
-	})
-
-	functionEventB := suite.CompileCreateFunctionEventOptions("function-event-b", functionToDeleteB.Meta.Name)
-	err = suite.Platform.CreateFunctionEvent(suite.Ctx, functionEventB)
-	suite.Require().NoError(err)
-	defer suite.Platform.DeleteFunctionEvent(suite.Ctx, &platform.DeleteFunctionEventOptions{ // nolint: errcheck
-		Meta: platform.FunctionEventMeta{
-			Name:      functionEventB.FunctionEventConfig.Meta.Name,
-			Namespace: suite.Namespace,
-		},
-	})
-
-	// try restrict - expect it to fail (project has sub resources)
-	err = suite.Platform.DeleteProject(suite.Ctx, &platform.DeleteProjectOptions{
-		Meta: platform.ProjectMeta{
-			Name:      projectToDeleteConfig.Meta.Name,
-			Namespace: suite.Namespace,
-		},
-		Strategy: platform.DeleteProjectStrategyRestricted,
-	})
-	suite.Require().Error(err)
-
-	// try cascading - expect it succeed
-	err = suite.Platform.DeleteProject(suite.Ctx, &platform.DeleteProjectOptions{
-		Meta: platform.ProjectMeta{
-			Name:      projectToDeleteConfig.Meta.Name,
-			Namespace: suite.Namespace,
-		},
-		Strategy:                           platform.DeleteProjectStrategyCascading,
-		WaitForResourcesDeletionCompletion: true,
-		WaitForResourcesDeletionCompletionDuration: 3 * time.Minute,
-	})
-	suite.Require().NoError(err)
-
-	// assertion - project should be deleted
-	projects, err := suite.Platform.GetProjects(suite.Ctx, &platform.GetProjectsOptions{
-		Meta: platform.ProjectMeta{
-			Name:      projectToDeleteConfig.Meta.Name,
-			Namespace: suite.Namespace,
-		},
-	})
-	suite.Require().NoError(err)
-	suite.Require().Len(projects, 0)
-
-	suite.Logger.InfoWith("Ensuring resources were removed (deletion is being executed in background")
-
-	// ensure api gateway deleted
-	apiGateways, err := suite.Platform.GetAPIGateways(suite.Ctx, &platform.GetAPIGatewaysOptions{
-		Name:      createAPIGatewayOptions.APIGatewayConfig.Meta.Name,
-		Namespace: suite.Namespace,
-	})
-	suite.Require().NoError(err)
-	suite.Require().Len(apiGateways, 0, "Some api gateways were not removed")
-
-	// ensure functions were deleted successfully
-	for _, functionName := range []string{
-		functionToDeleteA.Meta.Name,
-		functionToDeleteB.Meta.Name,
-	} {
-		functions, err := suite.Platform.GetFunctions(suite.Ctx, &platform.GetFunctionsOptions{
-			Name:      functionName,
-			Namespace: suite.Namespace,
-		})
-		suite.Require().NoError(err)
-		suite.Require().Len(functions, 0, "Some functions were not removed")
-	}
-
-	// ensure function events were deleted successfully
-	for _, functionEventName := range []string{
-		functionEventA.FunctionEventConfig.Meta.Name,
-		functionEventB.FunctionEventConfig.Meta.Name,
-	} {
-		functionEvents, err := suite.Platform.GetFunctionEvents(suite.Ctx, &platform.GetFunctionEventsOptions{
-			Meta: platform.FunctionEventMeta{
-				Name:      functionEventName,
-				Namespace: suite.Namespace,
-			},
-		})
-		suite.Require().NoError(err)
-		suite.Require().Len(functionEvents, 0, "Some function events were not removed")
-	}
-}
-
-func TestPlatformTestSuite(t *testing.T) {
+func TestFunctionTestSuite(t *testing.T) {
 	if testing.Short() {
 		return
 	}
 	suite.Run(t, new(DeployFunctionTestSuite))
 	suite.Run(t, new(UpdateFunctionTestSuite))
 	suite.Run(t, new(DeleteFunctionTestSuite))
-	suite.Run(t, new(DeployAPIGatewayTestSuite))
-	suite.Run(t, new(ProjectTestSuite))
 }
